refactor(agent-sessions): make scope updates atomic
