+++ packages/adapters-codex-app-server/src/codex-app-server-adapter.lifecycle.test.ts
@@ -4,7 +4,6 @@ import {
   codexSessionRef,
   codexSessionRuntimeRef,
   codexUserMessageInput,
-  createAdapterWithTransport,
   createHarness,
   defaultCodexEffectivePolicy,
   defaultCodexRuntimeConfig,
@@ -15,23 +14,12 @@ import {
 import { codexSandboxPolicy } from "./codex-session-policy";
 import { CodexAppServerAdapter } from "./index";
 
-class NameFailingTransport extends RecordingTransport {
-  async request<Response>(
-    request: Parameters<RecordingTransport["request"]>[0],
-  ): Promise<Response> {
-    if (request.method === "thread/name/set") {
-      this.calls.push(request);
-      throw new Error("name failed");
-    }
-    return super.request<Response>(request);
-  }
-}
-
 const localSessions = (
   adapter: CodexAppServerAdapter,
 ): { has(externalSessionId: string): boolean } =>
   (adapter as unknown as { localSessions: { has(externalSessionId: string): boolean } })
     .localSessions;
+
 const expectedThreadPolicy = {
   approvalPolicy: "on-request",
   approvalsReviewer: "user",
@@ -140,292 +128,6 @@ describe("CodexAppServerAdapter lifecycle", () => {
         name: "BUILD task-1",
       },
     });
-  });
-
-  test("applies repository policy across start, send, fork, resume, and history", async () => {
-    const sessionScope = { kind: "repository" } as const;
-    const runtimePolicy = { kind: "codex" as const, policy: defaultCodexEffectivePolicy() };
-    const model = { providerId: "openai", modelId: "gpt-5", variant: "medium" } as const;
-    const { adapter, transports } = createHarness();
-
-    const started = await adapter.startSession({
-      repoPath: "/repo",
-      runtimeKind: "codex",
-      workingDirectory: "/repo",
-      sessionScope,
-      runtimePolicy,
-      systemPrompt: "Use the repo rules.",
-      model,
-    });
-    await adapter.sendUserMessage(
-      codexUserMessageInput({
-        externalSessionId: started.externalSessionId,
-        sessionScope,
-        runtimePolicy,
-        parts: [{ kind: "text", text: "Continue" }],
-      }),
-    );
-    await adapter.loadSessionHistory(
-      codexSessionRuntimeRef(started.externalSessionId, { sessionScope, runtimePolicy }),
-    );
-    const forked = await adapter.forkSession({
-      repoPath: "/repo",
-      runtimeKind: "codex",
-      workingDirectory: "/repo",
-      parentExternalSessionId: started.externalSessionId,
-      sessionScope,
-      runtimePolicy,
-      systemPrompt: "Use the repo rules.",
-      model,
-    });
-    const resumed = await adapter.resumeSession({
-      repoPath: "/repo",
-      runtimeKind: "codex",
-      workingDirectory: "/repo",
-      externalSessionId: "thread-resume",
-      sessionScope,
-      runtimePolicy,
-      systemPrompt: "Use the repo rules.",
-      model,
-    });
-
-    expect(started).toMatchObject({
-      title: "Repository session",
-      sessionAssociation: sessionScope,
-      workingDirectory: "/repo",
-    });
-    expect(forked).toMatchObject({
-      title: "Repository session",
-      sessionAssociation: sessionScope,
-    });
-    expect(resumed).toMatchObject({
-      title: "Repository session",
-      sessionAssociation: sessionScope,
-    });
-    const calls = transports.get("runtime-live")?.calls ?? [];
-    expect(calls.filter((call) => call.method === "thread/name/set")).toEqual([
-      {
-        method: "thread/name/set",
-        params: { threadId: started.externalSessionId, name: "Repository session" },
-      },
-      {
-        method: "thread/name/set",
-        params: { threadId: forked.externalSessionId, name: "Repository session" },
-      },
-      {
-        method: "thread/name/set",
-        params: { threadId: resumed.externalSessionId, name: "Repository session" },
-      },
-    ]);
-    for (const call of calls.filter((candidate) =>
-      ["thread/start", "thread/fork", "thread/resume"].includes(candidate.method),
-    )) {
-      expect(call.params).toMatchObject({
-        cwd: "/repo",
-        ...expectedThreadPolicy,
-        config: { "mcp_servers.openducktor.enabled": false },
-      });
-    }
-  });
-
-  test("rejects a different workflow task or role for a bound session", async () => {
-    const { adapter, transports } = createHarness();
-    const started = await adapter.startSession({
-      repoPath: "/repo",
-      runtimeKind: "codex",
-      workingDirectory: "/repo",
-      sessionScope: { kind: "workflow", taskId: "task-1", role: "spec" },
-      runtimePolicy: { kind: "codex", policy: defaultCodexEffectivePolicy() },
-      systemPrompt: "Use the repo rules.",
-      model: { providerId: "openai", modelId: "gpt-5", variant: "medium" },
-    });
-    const callCount = transports.get("runtime-live")?.calls.length;
-
-    await expect(
-      adapter.resumeSession({
-        repoPath: "/repo",
-        runtimeKind: "codex",
-        workingDirectory: "/repo",
-        externalSessionId: started.externalSessionId,
-        sessionScope: { kind: "workflow", taskId: "task-2", role: "build" },
-        runtimePolicy: { kind: "codex", policy: defaultCodexEffectivePolicy() },
-        systemPrompt: "Use the repo rules.",
-        model: { providerId: "openai", modelId: "gpt-5", variant: "medium" },
-      }),
-    ).rejects.toThrow("registered workflow scope for task 'task-1' and role 'spec'");
-
-    await expect(
-      adapter.sendUserMessage(
-        codexUserMessageInput({
-          externalSessionId: started.externalSessionId,
-          sessionScope: { kind: "workflow", taskId: "task-1", role: "build" },
-          parts: [{ kind: "text", text: "Continue" }],
-        }),
-      ),
-    ).rejects.toThrow("requested workflow scope for task 'task-1' and role 'build'");
-    await expect(
-      adapter.loadSessionHistory(
-        codexSessionRuntimeRef(started.externalSessionId, {
-          sessionScope: { kind: "workflow", taskId: "task-2", role: "spec" },
-        }),
-      ),
-    ).rejects.toThrow("requested workflow scope for task 'task-2' and role 'spec'");
-    await expect(
-      adapter.loadSessionContextUsage(
-        codexSessionRuntimeRef(started.externalSessionId, {
-          sessionScope: { kind: "workflow", taskId: "task-1", role: "qa" },
-        }),
-      ),
-    ).rejects.toThrow("requested workflow scope for task 'task-1' and role 'qa'");
-    expect(transports.get("runtime-live")?.calls).toHaveLength(callCount ?? 0);
-  });
-
-  test("allows an explicit workflow role change only when forking", async () => {
-    const { adapter } = createHarness();
-    const started = await adapter.startSession({
-      repoPath: "/repo",
-      runtimeKind: "codex",
-      workingDirectory: "/repo",
-      sessionScope: { kind: "workflow", taskId: "task-1", role: "spec" },
-      runtimePolicy: { kind: "codex", policy: defaultCodexEffectivePolicy() },
-      systemPrompt: "Use the repo rules.",
-      model: { providerId: "openai", modelId: "gpt-5", variant: "medium" },
-    });
-
-    await expect(
-      adapter.forkSession({
-        repoPath: "/repo",
-        runtimeKind: "codex",
-        workingDirectory: "/repo",
-        parentExternalSessionId: started.externalSessionId,
-        sessionScope: { kind: "workflow", taskId: "task-1", role: "build" },
-        runtimePolicy: { kind: "codex", policy: defaultCodexEffectivePolicy() },
-        systemPrompt: "Use the repo rules.",
-        model: { providerId: "openai", modelId: "gpt-5", variant: "medium" },
-      }),
-    ).resolves.toMatchObject({
-      title: "BUILD task-1",
-      sessionAssociation: { kind: "workflow", taskId: "task-1", role: "build" },
-    });
-  });
-
-  test("rejects a workflow scope for a repository-bound session", async () => {
-    const { adapter, transports } = createHarness();
-    const started = await adapter.startSession({
-      repoPath: "/repo",
-      runtimeKind: "codex",
-      workingDirectory: "/repo",
-      sessionScope: { kind: "repository" },
-      runtimePolicy: { kind: "codex", policy: defaultCodexEffectivePolicy() },
-      systemPrompt: "Use the repo rules.",
-      model: { providerId: "openai", modelId: "gpt-5", variant: "medium" },
-    });
-    const callCount = transports.get("runtime-live")?.calls.length;
-
-    await expect(
-      adapter.resumeSession({
-        repoPath: "/repo",
-        runtimeKind: "codex",
-        workingDirectory: "/repo",
-        externalSessionId: started.externalSessionId,
-        sessionScope: { kind: "workflow", taskId: "task-1", role: "build" },
-        runtimePolicy: { kind: "codex", policy: defaultCodexEffectivePolicy() },
-        systemPrompt: "Use the repo rules.",
-        model: { providerId: "openai", modelId: "gpt-5", variant: "medium" },
-      }),
-    ).rejects.toThrow("registered repository scope does not match the requested workflow scope");
-
-    await expect(
-      adapter.sendUserMessage(
-        codexUserMessageInput({
-          externalSessionId: started.externalSessionId,
-          sessionScope: { kind: "workflow", taskId: "task-1", role: "build" },
-          parts: [{ kind: "text", text: "Continue" }],
-        }),
-      ),
-    ).rejects.toThrow("registered repository scope does not match the requested workflow scope");
-    expect(transports.get("runtime-live")?.calls).toHaveLength(callCount ?? 0);
-  });
-
-  test("rejects a start without session scope before runtime side effects", async () => {
-    const { adapter, requireRepoRuntime, transportFactory } = createHarness();
-
-    await expect(
-      adapter.startSession({
-        repoPath: "/repo",
-        runtimeKind: "codex",
-        workingDirectory: "/repo",
-        runtimePolicy: { kind: "codex", policy: defaultCodexEffectivePolicy() },
-        systemPrompt: "Use the repo rules.",
-        model: { providerId: "openai", modelId: "gpt-5", variant: "medium" },
-      } as never),
-    ).rejects.toThrow("Cannot start Codex session without session context.");
-    expect(requireRepoRuntime).toHaveBeenCalledTimes(0);
-    expect(transportFactory).toHaveBeenCalledTimes(0);
-  });
-
-  test("keeps started sessions addressable when thread naming fails", async () => {
-    const transport = new NameFailingTransport("runtime-live", false);
-    const adapter = createAdapterWithTransport(transport);
-
-    await expect(
-      adapter.startSession({
-        repoPath: "/repo",
-        runtimeKind: "codex",
-        workingDirectory: "/repo",
-        sessionScope: { kind: "workflow", taskId: "task-1", role: "build" },
-        runtimePolicy: { kind: "codex", policy: defaultCodexEffectivePolicy() },
-        systemPrompt: "Use the repo rules.",
-        model: { providerId: "openai", modelId: "gpt-5", variant: "medium" },
-      }),
-    ).rejects.toThrow("name failed");
-
-    expect(localSessions(adapter).has("thread/start-runtime-live")).toBe(true);
-    expect(transport.calls.map((call) => call.method)).toEqual([
-      "model/list",
-      "thread/start",
-      "thread/name/set",
-    ]);
-  });
-
-  test("keeps resumed and history-restored repository sessions when thread naming fails", async () => {
-    const repositoryScope = { kind: "repository" } as const;
-    const resumedTransport = new NameFailingTransport("runtime-live", false);
-    const resumedAdapter = createAdapterWithTransport(resumedTransport);
-
-    await expect(
-      resumedAdapter.resumeSession({
-        repoPath: "/repo",
-        runtimeKind: "codex",
-        workingDirectory: "/repo",
-        externalSessionId: "thread-resume",
-        sessionScope: repositoryScope,
-        runtimePolicy: { kind: "codex", policy: defaultCodexEffectivePolicy() },
-        systemPrompt: "Use the repo rules.",
-        model: { providerId: "openai", modelId: "gpt-5", variant: "medium" },
-      }),
-    ).rejects.toThrow("name failed");
-    expect(localSessions(resumedAdapter).has("thread-resume")).toBe(true);
-    expect(resumedTransport.calls.find((call) => call.method === "thread/resume")?.params).toEqual(
-      expect.objectContaining({ config: { "mcp_servers.openducktor.enabled": false } }),
-    );
-
-    const restoredTransport = new NameFailingTransport("runtime-live", false);
-    const restoredAdapter = createAdapterWithTransport(restoredTransport);
-    await expect(
-      restoredAdapter.sendUserMessage(
-        codexUserMessageInput({
-          externalSessionId: "thread-history",
-          sessionScope: repositoryScope,
-          runtimePolicy: { kind: "codex", policy: defaultCodexEffectivePolicy() },
-          parts: [{ kind: "text", text: "Continue" }],
-        }),
-      ),
-    ).rejects.toThrow("name failed");
-    expect(localSessions(restoredAdapter).has("thread-history")).toBe(true);
-    expect(restoredTransport.calls.find((call) => call.method === "thread/resume")?.params).toEqual(
-      expect.objectContaining({ config: { "mcp_servers.openducktor.enabled": false } }),
-    );
   });
 
   test("caches Codex model lists per runtime", async () => {
@@ -1043,33 +745,6 @@ describe("CodexAppServerAdapter lifecycle", () => {
     await expect(
       adapter.listAvailableModels({ repoPath: "/repo", runtimeKind: "codex" }),
     ).rejects.toThrow("No live repo runtime found for repo '/repo' and runtime 'codex'.");
-  });
-
-  test("fails repository history restoration on a missing bound Codex route", async () => {
-    const requireRepoRuntime = mock(async () => ({
-      ...makeRuntimeSummary("runtime-wrong-route"),
-      runtimeRoute: { type: "local_http" as const, endpoint: "http://127.0.0.1:43123" },
-    }));
-    const transportFactory = mock(() => {
-      throw new Error("transportFactory should not be called");
-    });
-    const adapter = new CodexAppServerAdapter({
-      repoRuntimeResolver: { requireRepoRuntime },
-      transportFactory,
-    });
-
-    await expect(
-      adapter.loadSessionHistory(
-        codexSessionRuntimeRef("repository-history", {
-          sessionScope: { kind: "repository" },
-          runtimePolicy: { kind: "codex", policy: defaultCodexEffectivePolicy() },
-        }),
-      ),
-    ).rejects.toThrow(
-      "runtime 'runtime-wrong-route' is missing required route contract 'stdio' for repository session 'repository-history'",
-    );
-    expect(requireRepoRuntime).toHaveBeenCalledTimes(1);
-    expect(transportFactory).toHaveBeenCalledTimes(0);
   });
 
   test("starts a session in the requested build worktree through the repo runtime", async () => {
